clean up utility functions
