+++ donkeycar/tests/test_keras.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 import pytest
-from donkeycar.parts.keras import KerasPilot, KerasCategorical, KerasIMU
-from donkeycar.parts.keras import default_categorical, default_imu, default_n_linear
+from donkeycar.parts.keras import KerasPilot, KerasCategorical
+from donkeycar.parts.keras import default_categorical, default_n_linear
 # content of ./test_smtpsimple.py
 
 @pytest.fixture
@@ -24,13 +24,5 @@ def test_categorical_with_model():
 
 def test_def_n_linear_model():
     model = default_n_linear(10)
-    assert model is not None   
-
-def test_imu_model():
-    model = default_imu(2, 7)
     assert model is not None
-
-def test_KerasIMU():
-    k = KerasIMU(num_outputs=2, num_imu_inputs=7)
-    assert k.model is not None
