+++ donkeycar/tests/test_scripts.py
@@ -1,4 +1,4 @@
-from donkeycar import utils
+from donkeycar import util
 import pytest
 
 
@@ -21,14 +21,14 @@ def cardir(tmpdir):
 
 def test_createcar(cardir):
     cmd = ['donkey', 'createcar', '--path', cardir]
-    out, err, proc_id = utils.run_shell_command(cmd)
+    out, err, proc_id = util.proc.run_shell_command(cmd)
     assert is_error(err) is False
 
 def test_drivesim(cardir):
     cmd = ['donkey', 'createcar', '--path', cardir ,'--template', 'square']
-    out, err, proc_id = utils.run_shell_command(cmd, timeout=10)
+    out, err, proc_id = util.proc.run_shell_command(cmd, timeout=10)
     cmd = ['python', 'manage.py', 'drive']
-    out, err, proc_id = utils.run_shell_command(cmd, cwd = cardir)
+    out, err, proc_id = util.proc.run_shell_command(cmd, cwd = cardir)
     print(err)
 
     if is_error(err) is True:
@@ -38,7 +38,7 @@ def test_drivesim(cardir):
 
 def test_bad_command_fails():
     cmd = ['donkey', 'not a comand']
-    out, err, proc_id = utils.run_shell_command(cmd)
+    out, err, proc_id = util.proc.run_shell_command(cmd)
     print(err)
     print(out)
     assert is_error(err) is True
