+++ donkeycar/util/__init__.py
@@ -1,0 +1,6 @@
+from . import (proc,
+               data,
+               files,
+               img,
+               times,
+               web)
