+++ donkeycar/util/proc.py
@@ -1,0 +1,41 @@
+"""
+Functions to simplify working with processes.
+"""
+import subprocess
+
+
+
+def run_shell_command(cmd, cwd=None, timeout=15):
+    proc = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, cwd=cwd)
+    out = []
+    err = []
+
+    try:
+        proc.wait(timeout=timeout)
+    except subprocess.TimeoutExpired:
+        kill(proc.pid)
+
+    for line in proc.stdout.readlines():
+        out.append(line.decode())
+
+    for line in proc.stderr.readlines():
+        err.append(line)
+    return out, err, proc.pid
+
+'''
+def kill(proc_pid):
+    process = psutil.Process(proc_pid)
+    for proc in process.children(recursive=True):
+        proc.kill()
+    process.kill()
+'''
+import signal
+
+def kill(proc_id):
+    os.kill(proc_id, signal.SIGINT)
+
+
+
+
+def eprint(*args, **kwargs):
+    print(*args, file=sys.stderr, **kwargs)
